Remove old version of the spraylist and linden sources which are not used
This commit replaces the old version of spraylist sources with an
updated version. It also removes the separate folder for the Linden
priority queue sources and makes sure that the version of the Linden
priority queue that is located in the spraylist folder is used
instead. The version of Linden's priority queue that is located in the
spraylist folder is modified by the spraylist authors to allow
duplicated keys.